requirements.txt
fastf1
pandas
pyyaml
tqdm
pyarrow
